Add `Decimal.sum/1`
We don't have any other functions that operate on enumerables, so if
this is accepted we're probably opening doors for even more.
That being said, it's so common to do something like this that I think
it would be good to have this shortcut rather than having people to use
`Enum.reduce`.